colocate program serializers with their view files
